+++ src/app/admin/early-bird/early-bird.component.html
@@ -1,0 +1,119 @@
+<br>
+<br>
+<!-- EarlyBird ------------------------------------------------------------------------------------------------------------------------------------------------------------>
+<!-- EarlyBird button -->
+<div class="row align-items-center mb-3">
+    <div class="col-md-6">
+      <h2>Early Bird</h2>
+    </div>
+    <div class="col-md-6 text-end">
+      <button class="btn btn-success btn-sm" (click)="openAddEarlyBirdModal()">
+        <i class="fas fa-plus"></i> Add
+      </button>
+    </div>
+  </div>
+  
+  <!--EarlyBird table -->
+  <table class="table table-striped">
+    <thead>
+      <tr>
+        <!-- <th scope="col">ID</th> -->
+        <th scope="col">Percentage</th>
+        <th scope="col">Limit</th>
+      </tr>
+    </thead>
+    <tbody>
+      <tr *ngFor="let earlyBird of earlyBirds">
+        <!-- <td>{{ earlyBird.earlyBirdID }}</td> -->
+        <td>{{ earlyBird.percentage }}</td>
+        <td>{{ earlyBird.limit }}</td>
+        <td>
+          <button class="btn btn-primary"
+            (click)="earlyBird.earlyBirdID !== undefined && openEditEarlyBirdModal(earlyBird.earlyBirdID)"
+            style="margin-right: 4px;">
+            <i class="fas fa-pencil-alt"></i>
+          </button>
+  
+          <button class="btn btn-danger" (click)="earlyBird.earlyBirdID !== undefined && openDeleteEarlyBirdModal(earlyBird)">
+            <i class="fas fa-trash-alt"></i>
+          </button>
+        </td>
+      </tr>
+    </tbody>
+  </table>
+
+  <div class="text-center" *ngIf="earlyBirds.length === 0">
+    <h3>No early bird options currently in your system.</h3>
+  </div>
+  
+ <!-- Add/Edit EarlyBird Modal -->
+<div class="modal" tabindex="-1" [ngClass]="{'show': showEarlyBirdModal}" [ngStyle]="{'display': showEarlyBirdModal ? 'block' : 'none'}">
+  <div class="modal-dialog">
+    <div class="modal-content">
+      <div class="modal-header">
+        <h5 class="modal-title">{{editingEarlyBird ? 'Edit EarlyBird' : 'Add EarlyBird'}}</h5>
+        <button type="button" class="btn-close" aria-label="Close" (click)="closeEarlyBirdModal()"></button>
+      </div>
+      <div class="modal-body">
+        <form #earlyBirdForm="ngForm" (ngSubmit)="submitEarlyBirdForm(earlyBirdForm)">
+
+          <div class="mb-3">
+            <label for="percentage" class="form-label">Percentage</label>
+            <input type="number" class="form-control" id="percentage" [(ngModel)]="currentEarlyBird.percentage"
+              name="percentage" percentage required min="1" max="100" #percentage="ngModel">
+            <div *ngIf="percentage.invalid && percentage.touched" class="alert alert-danger">
+              <div *ngIf="percentage.errors?.['required']">Percentage is required.</div>
+              <div *ngIf="percentage.value === 0">Percentage cannot be 0.</div>
+             
+              <div *ngIf="percentage.errors?.['max']">Percentage cannot be more than 100.</div>
+             
+             
+            </div>
+          </div>
+
+          <div class="mb-3">
+            <label for="limit" class="form-label">Limit</label>
+            <input type="number" class="form-control" id="limit" [(ngModel)]="currentEarlyBird.limit" name="limit"
+              limit required min="1" max="998" #limit="ngModel">
+            <div *ngIf="limit.invalid && limit.touched" class="alert alert-danger">
+              <div *ngIf="limit.errors?.['required']">Limit is required.</div>
+              <div *ngIf="limit.value === 999">Limit cannot be 999.</div>
+              
+            </div>
+          </div>
+
+          <div class="modal-footer">
+            <button type="button" class="btn btn-secondary" (click)="closeEarlyBirdModal()">Close</button>
+            <button type="submit" class="btn btn-primary" [disabled]="earlyBirdForm.invalid">{{editingEarlyBird ? 'Save Changes' : 'Save'}}</button>
+          </div>
+        </form>
+      </div>
+    </div>
+  </div>
+</div>
+
+  <!-- Delete EarlyBird modal -->
+<div class="modal" tabindex="-1" [ngClass]="{'show': showDeleteEarlyBirdModal}"
+[ngStyle]="{'display': showDeleteEarlyBirdModal ? 'block' : 'none'}">
+<div class="modal-dialog">
+  <div class="modal-content">
+    <div class="modal-header">
+      <h5 class="modal-title">Delete Confirmation</h5>
+      <button type="button" class="btn-close" aria-label="Close" (click)="closeDeleteModal()"></button>
+    </div>
+    <div class="modal-body">
+      Are you sure you want to delete this Early Bird option?
+      <p></p>
+      <p>ID: {{ earlyBirdToDeleteDetails?.earlyBirdID }}</p>
+      <p>Percentage: {{ earlyBirdToDeleteDetails?.percentage }}</p>
+      <p>Limit: {{ earlyBirdToDeleteDetails?.limit }}</p>
+    </div>
+    <div class="modal-footer">
+      <button type="button" class="btn btn-secondary" (click)="closeDeleteModal()">No</button>
+      <button type="button" class="btn btn-primary" (click)="deleteEarlyBird()">Yes</button>
+    </div>
+  </div>
+</div>
+</div>
+<!-- EarlyBird ---------------------------------------------------------------------------------------------------------------------------------------------------------->
+
