+++ src/app/admin/early-bird/early-bird.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { EarlyBirdComponent } from './early-bird.component';
+
+describe('EarlyBirdComponent', () => {
+  let component: EarlyBirdComponent;
+  let fixture: ComponentFixture<EarlyBirdComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [ EarlyBirdComponent ]
+    })
+    .compileComponents();
+
+    fixture = TestBed.createComponent(EarlyBirdComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
