+++ src/app/admin/sidenav/sidenav.component.html
@@ -38,11 +38,24 @@
         <span>Inventory</span>
       </a>
     </li>
-    <li class="nav-item">
-      <a class="nav-link" routerLink="/event" routerLinkActive="active">
+    <li class="nav-item dropdown" (click)="showEventSubnav = !showEventSubnav">
+      <a class="nav-link">
         <i class="bi bi-calendar-event me-2"></i>
         <span>Event</span>
+        <i class="bi bi-chevron-down" style="margin-left: 15px;"></i>
       </a>
+      <ul *ngIf="showEventSubnav" class="nav flex-column ml-3 sub-nav">
+        <li class="nav-item">
+          <a class="nav-link" routerLink="/event">
+            <span>Event</span>
+          </a>
+        </li>
+        <li class="nav-item">
+          <a class="nav-link" routerLink="/earlybird">
+            <span>Early Bird</span>
+          </a>
+        </li>
+      </ul>
     </li>
     <li class="nav-item">
       <a class="nav-link" routerLink="/faq" routerLinkActive="active">
@@ -82,3 +95,4 @@
     </li>
   </ul>
 </div>
+
