+++ src/app/app-routing.module.ts
@@ -12,6 +12,7 @@ import { EmployeeComponent } from './admin/employee/employee.component';
 import { VarietalComponent } from './admin/varietal/varietal.component';
 import { TypeComponent } from './admin/type/type.component';
 import { DiscountComponent } from './admin/discount/discount.component';
+import { EarlyBirdComponent } from './admin/early-bird/early-bird.component';
 
 
 
@@ -28,7 +29,8 @@ const routes: Routes = [
   { path: 'employees', component: EmployeeComponent},
   {path: 'varietal', component: VarietalComponent},
   {path: 'type', component: TypeComponent},
-  {path: 'discount', component: DiscountComponent}
+  {path: 'discount', component: DiscountComponent},
+  {path: 'earlybird', component: EarlyBirdComponent}
 
 ];
